Threshold added

+++ DCT/DCTLib/Matrix.hpp
@@ -9,7 +9,7 @@ using std::array;
 using Matrix8x8ui8 = array<array<uint8_t, 8>, 8>;
 using Matrix8x8i8 = array<array<int8_t, 8>, 8>;
 
-Matrix8x8ui8 BuildQuantizationMatrix(const float quality);
+Matrix8x8ui8 BuildQuantizationMatrix(const uint8_t quality);
 Matrix8x8i8 Encode(const Matrix8x8ui8 &inputMat, const Matrix8x8ui8 &q);
 Matrix8x8ui8 Decode(const Matrix8x8i8 &inputMat, const Matrix8x8ui8 &q);
